Add singledispatch benchmarks
Add some benchmarks that measure the performance of calling a function
that dispatches to an implementation using singledispatch.

This differentiates between 2 types of situations that we benchmark:
dynamic dispatch and static dispatch. Dynamic dispatch occurs when we
don't know the type of the dispatch argument, in which case we want to
make the type check as fast as possible. However, if we do know the type
of the dispatch argument at compile time, we can skip the runtime type
check completely and use the correct implementation directly.