Add mypy self check benchmark
This compiles the versions of mypy being tested, using the tested version of
mypyc, and measures the time needed to type check a specific vendored version
of mypy.

Also install pinned versions of PEP 561 dependencies for consistent results.

The vendored copy of mypy doesn't include mypyc or typeshed. Typeshed will
be used from the tested mypy.