Add memory allocation microbenchmarks (#45)
Allocating and freeing instances is often a major bottleneck. Add
microbenchmarks that model simplified versions of a few different
scenarios. Add benchmarks for both short-lived and long-lived objects.
The latter will likely spend more time in the cyclic GC. Also
separately benchmark instances which don't need to participate in the
cyclic GC. Currently all native classes participate in GC, but we
could opt out classes from it whenever possible.

We have other benchmarks that also heavily allocate objects, but it can 
still be useful to have dedicated microbenchmarks for this.